feat(org-workspace): SSR prefetch → HydrationBoundary on all 4 pages (Part of #966)
Closes the last data-flow gap from the audit: org-workspace was the only
dashboard tree fetching entirely client-side (0 SSR prefetch). All four pages
now match the org flagship pattern — a thin server page.tsx prefetches into a
QueryClient and hands a dehydrated cache to a client *PageClient via
HydrationBoundary, so the first paint has data instead of a spinner.

- New lib/data/org-workspace.ts: four DB-direct read functions
  (getOperatorOrganizations / getWorkspaceBillingRollup / getWorkspaceActivity
  / getWorkspaceSettings) taking an already-authenticated userId and returning
  byte-identical shapes to the route JSON bodies.
- The 4 GET routes (/api/organizations, org-workspace billing/activity/settings)
  keep their auth + IDOR guards and delegate the body to the lib fn — one code
  path for SSR and CSR, no drift. POST/PATCH handlers untouched; settings keeps
  its 404-when-absent via try/catch.
- Pages split into server page.tsx (await params → requireAuth → Promise.allSettled
  prefetch → HydrationBoundary) + client *PageClient.tsx (useQuery/useInfiniteQuery/
  useWorkspaceBilling unchanged, hydrated from cache). Query keys are byte-identical
  to the client's so hydration applies. Activity uses prefetchInfiniteQuery
  (initialPageParam: null). Billing roll-up is prefetched on both home and billing.

Correctness: Promise.allSettled means a failed prefetch silently degrades to the
existing client fetch (never breaks the page); activity.createdAt / settings.updatedAt
Prisma Dates are pre-stringified in the lib fns so the dehydrated cache matches the
client's string types (no hydration mismatch).

Monolith file decomposition (programs 2021 / moderation 1023) intentionally left
out — cosmetic churn with no functional value and real blind-refactor risk.

tsc clean; changed files lint-clean; 1498 jest tests pass.
